refactor: Organize lib/ into layer-based directory
- Separate library into logical layers of actions and validation
