+++ app/techniques/techniques-list.tsx
@@ -5,7 +5,7 @@ import { Technique } from "../generated/prisma/client";
 import { createColumns } from "./columns";
 import { DataTable } from "./data-table";
 import { NewTechniqueForm } from "./new-technique-form";
-import { deleteTechnique } from "@/lib/techniques";
+import { deleteTechnique } from "@/lib/actions/techniques";
 import { Button } from "@/components/ui/button";
 import {
   Collapsible,
